Extract `spacetimedb-memory-usage` & Nix `bindings` dep on `spacetimedb-data-structures` (#2919)
