+++ crates/bindings/tests/ui/tables.stderr
@@ -21,12 +21,12 @@ error[E0277]: the column type `Test` does not implement `SpacetimeType`
   = help: the following other types implement trait `SpacetimeType`:
             &T
             ()
+            AlgebraicType
             AlgebraicTypeRef
             Alpha
             Arc<T>
             ArrayType
             Box<T>
-            ColId
           and $N others
   = note: required for `Test` to implement `TableColumn`
 
@@ -40,12 +40,12 @@ error[E0277]: the trait bound `Test: SpacetimeType` is not satisfied
   = help: the following other types implement trait `SpacetimeType`:
             &T
             ()
+            AlgebraicType
             AlgebraicTypeRef
             Alpha
             Arc<T>
             ArrayType
             Box<T>
-            ColId
           and $N others
 
 error[E0277]: the trait bound `Test: Deserialize<'de>` is not satisfied
@@ -61,11 +61,11 @@ error[E0277]: the trait bound `Test: Deserialize<'de>` is not satisfied
             &'de [u8]
             &'de str
             ()
+            AlgebraicType
             AlgebraicTypeRef
             Alpha
             Arc<[T]>
             ArrayType
-            Box<T>
           and $N others
 note: required by a bound in `spacetimedb::spacetimedb_lib::de::SeqProductAccess::next_element`
  --> $WORKSPACE/crates/sats/src/de.rs
@@ -83,11 +83,11 @@ error[E0277]: the trait bound `Test: Deserialize<'_>` is not satisfied
             &'de [u8]
             &'de str
             ()
+            AlgebraicType
             AlgebraicTypeRef
             Alpha
             Arc<[T]>
             ArrayType
-            Box<T>
           and $N others
 note: required by a bound in `get_field_value`
  --> $WORKSPACE/crates/sats/src/de.rs
@@ -104,12 +104,12 @@ error[E0277]: the trait bound `Test: Serialize` is not satisfied
   = help: the following other types implement trait `Serialize`:
             &T
             ()
+            AlgebraicType
             AlgebraicTypeRef
+            AlgebraicValue
             Alpha
             Arc<T>
             ArrayType
-            ArrayValue
-            Bound<T>
           and $N others
 note: required by a bound in `spacetimedb::spacetimedb_lib::ser::SerializeNamedProduct::serialize_element`
  --> $WORKSPACE/crates/sats/src/ser.rs
@@ -135,7 +135,7 @@ error[E0277]: `&'a Alpha` cannot appear as an argument to an index filtering ope
              &bool
              &ethnum::int::I256
            and $N others
-note: required by a bound in `UniqueColumn::<Tbl, <Col as spacetimedb::table::Column>::ColType, Col>::find`
+note: required by a bound in `UniqueColumn::<Tbl, <Col as Column>::ColType, Col>::find`
   --> src/table.rs
    |
    |     pub fn find(&self, col_val: impl Borrow<Col::ColType>) -> Option<Tbl::Row>
