+++ crates/core/src/db/datastore/error.rs
@@ -1,15 +1,11 @@
 use super::system_tables::SystemTable;
 use enum_as_inner::EnumAsInner;
-use spacetimedb_lib::buffer::DecodeError;
-use spacetimedb_lib::{
-    db::{
-        error::LibError,
-        raw_def::{v9::RawSql, RawIndexDefV8},
-    },
-    AlgebraicType, AlgebraicValue, ProductValue,
-};
+use spacetimedb_lib::db::raw_def::{v9::RawSql, RawIndexDefV8};
 use spacetimedb_primitives::{ColId, ColList, IndexId, SequenceId, TableId};
+use spacetimedb_sats::buffer::DecodeError;
 use spacetimedb_sats::{product_value::InvalidFieldError, satn::Satn};
+use spacetimedb_sats::{AlgebraicType, AlgebraicValue, ProductValue};
+use spacetimedb_schema::def::error::LibError;
 use spacetimedb_snapshot::SnapshotError;
 use spacetimedb_table::{
     bflatn_to, read_column,
