+++ crates/core/src/error.rs
@@ -16,11 +16,11 @@ use thiserror::Error;
 use crate::client::ClientActorId;
 use crate::host::scheduler::ScheduleError;
 use spacetimedb_lib::buffer::DecodeError;
-use spacetimedb_lib::db::error::{LibError, RelationError, SchemaErrors};
-use spacetimedb_lib::relation::FieldName;
 use spacetimedb_primitives::*;
 use spacetimedb_sats::hash::Hash;
 use spacetimedb_sats::product_value::InvalidFieldError;
+use spacetimedb_schema::def::error::{LibError, RelationError, SchemaErrors};
+use spacetimedb_schema::relation::FieldName;
 use spacetimedb_vm::errors::{ErrorKind, ErrorLang, ErrorType, ErrorVm};
 use spacetimedb_vm::expr::Crud;
