+++ crates/core/src/sql/ast.rs
@@ -7,11 +7,11 @@ use spacetimedb_data_structures::map::{HashCollectionExt as _, IntMap};
 use spacetimedb_expr::check::SchemaView;
 use spacetimedb_expr::statement::compile_sql_stmt;
 use spacetimedb_lib::db::auth::StAccess;
-use spacetimedb_lib::db::error::RelationError;
 use spacetimedb_lib::identity::AuthCtx;
-use spacetimedb_lib::relation::{ColExpr, FieldName};
 use spacetimedb_primitives::{ColId, TableId};
 use spacetimedb_sats::{AlgebraicType, AlgebraicValue};
+use spacetimedb_schema::def::error::RelationError;
+use spacetimedb_schema::relation::{ColExpr, FieldName};
 use spacetimedb_schema::schema::{ColumnSchema, TableSchema};
 use spacetimedb_vm::errors::ErrorVm;
 use spacetimedb_vm::expr::{Expr, FieldExpr, FieldOp};
