+++ crates/core/src/sql/compiler.rs
@@ -7,8 +7,8 @@ use crate::error::{DBError, PlanError};
 use core::ops::Deref;
 use spacetimedb_data_structures::map::IntMap;
 use spacetimedb_lib::identity::AuthCtx;
-use spacetimedb_lib::relation::{self, ColExpr, DbTable, FieldName, Header};
 use spacetimedb_primitives::ColId;
+use spacetimedb_schema::relation::{self, ColExpr, DbTable, FieldName, Header};
 use spacetimedb_schema::schema::TableSchema;
 use spacetimedb_vm::expr::{CrudExpr, Expr, FieldExpr, QueryExpr, SourceExpr};
 use spacetimedb_vm::operator::OpCmp;
