+++ crates/core/src/vm.rs
@@ -12,9 +12,9 @@ use core::ops::{Bound, RangeBounds};
 use itertools::Itertools;
 use spacetimedb_data_structures::map::IntMap;
 use spacetimedb_lib::identity::AuthCtx;
-use spacetimedb_lib::relation::{ColExpr, DbTable};
 use spacetimedb_primitives::*;
 use spacetimedb_sats::{AlgebraicValue, ProductValue};
+use spacetimedb_schema::relation::{ColExpr, DbTable};
 use spacetimedb_table::static_assert_size;
 use spacetimedb_table::table::RowRef;
 use spacetimedb_vm::errors::ErrorVm;
@@ -652,9 +652,9 @@ pub(crate) mod tests {
     use pretty_assertions::assert_eq;
     use spacetimedb_lib::db::auth::{StAccess, StTableType};
     use spacetimedb_lib::error::ResultTest;
-    use spacetimedb_lib::relation::{FieldName, Header};
     use spacetimedb_sats::{product, AlgebraicType, ProductType, ProductValue};
     use spacetimedb_schema::def::{BTreeAlgorithm, IndexAlgorithm};
+    use spacetimedb_schema::relation::{FieldName, Header};
     use spacetimedb_schema::schema::{ColumnSchema, IndexSchema, TableSchema};
     use spacetimedb_vm::eval::run_ast;
     use spacetimedb_vm::eval::test_helpers::{mem_table, mem_table_one_u64, scalar};
