+++ crates/lib/src/lib.rs
@@ -15,7 +15,6 @@ pub mod identity;
 pub mod metrics;
 pub mod operator;
 pub mod query;
-pub mod relation;
 pub mod scheduler;
 pub mod st_var;
 pub mod version;
