+++ crates/memory-usage/README.md
@@ -1,0 +1,3 @@
+> ⚠️ **Internal Crate** ⚠️
+>
+> This crate is intended for internal use only. It is **not** stable and may change without notice.
