+++ crates/primitives/Cargo.toml
@@ -5,11 +5,15 @@ edition.workspace = true
 license-file = "LICENSE"
 description = "Primitives such as TableId and ColumnIndexAttribute"
 
+[features]
+memory-usage = ["dep:spacetimedb-memory-usage"]
+
 [dependencies]
 bitflags.workspace = true
 either.workspace = true
 nohash-hasher.workspace = true
 itertools.workspace = true
+spacetimedb-memory-usage = { workspace = true, optional = true, default-features = false }
 
 [dev-dependencies]
 proptest.workspace = true
