+++ crates/schema/src/lib.rs
@@ -6,5 +6,6 @@ pub mod auto_migrate;
 pub mod def;
 pub mod error;
 pub mod identifier;
+pub mod relation;
 pub mod schema;
 pub mod type_for_generate;
