+++ crates/vm/src/eval.rs
@@ -96,9 +96,9 @@ pub mod test_helpers {
     use crate::relation::MemTable;
     use core::hash::BuildHasher as _;
     use spacetimedb_data_structures::map::DefaultHashBuilder;
-    use spacetimedb_lib::relation::{Column, FieldName, Header};
     use spacetimedb_primitives::TableId;
     use spacetimedb_sats::{product, AlgebraicType, AlgebraicValue, ProductType, ProductValue};
+    use spacetimedb_schema::relation::{Column, FieldName, Header};
     use std::sync::Arc;
 
     pub fn mem_table_without_table_name(mem: &MemTable) -> (&[Column], &[ProductValue]) {
@@ -185,11 +185,11 @@ pub mod tests {
     use crate::expr::{CrudExpr, Query, QueryExpr, SourceExpr, SourceSet};
     use crate::iterators::RelIter;
     use crate::relation::MemTable;
-    use spacetimedb_lib::db::error::RelationError;
     use spacetimedb_lib::operator::{OpCmp, OpLogic};
-    use spacetimedb_lib::relation::{FieldName, Header};
     use spacetimedb_primitives::ColId;
     use spacetimedb_sats::{product, AlgebraicType, ProductType};
+    use spacetimedb_schema::def::error::RelationError;
+    use spacetimedb_schema::relation::{FieldName, Header};
 
     /// From an original source of `result`s, applies `queries` and returns a final set of results.
     fn build_query<'a, const N: usize>(
